minor cleanups. fixed footer_buttons overflow in eventpage, datetime pickers
